+++ extensions/gitlab/src/components/branch.tsx
@@ -1,9 +1,15 @@
-import { ActionPanel, List, OpenInBrowserAction, Image, Color, showToast, ToastStyle } from "@raycast/api";
+import { ActionPanel, List, Image, Color, showToast, Toast } from "@raycast/api";
 import { useEffect, useState } from "react";
 import { Project } from "../gitlabapi";
 import { gitlab } from "../common";
 import { GitLabIcons } from "../icons";
-import { CreateMRAction } from "./branch_actions";
+import { CreateMRAction, ShowBranchCommitsAction } from "./branch_actions";
+import { GitLabOpenInBrowserAction } from "./actions";
+import { useCommitStatus } from "./commits/utils";
+import { getCIJobStatusIcon } from "./jobs";
+import { ensureCleanAccessories } from "../utils";
+
+/* eslint-disable @typescript-eslint/no-explicit-any,@typescript-eslint/explicit-module-boundary-types */
 
 function getIcon(merged: boolean): Image {
   if (merged) {
@@ -16,6 +22,7 @@ function getIcon(merged: boolean): Image {
 export function BranchListItem(props: { branch: any; project: Project }) {
   const branch = props.branch;
   const icon = getIcon(branch.merged as boolean);
+  const project = props.project;
   const states = [];
   if (branch.default) {
     states.push("[default]");
@@ -23,16 +30,21 @@ export function BranchListItem(props: { branch: any; project: Project }) {
   if (branch.protected) {
     states.push("[protected]");
   }
+  const { commitStatus } = useCommitStatus(project.id, branch?.commit?.id);
+  const statusIcon = commitStatus ? getCIJobStatusIcon(commitStatus.status) : undefined;
+
   return (
     <List.Item
       id={branch.id}
       title={branch.name}
       subtitle={states.join(" ")}
       icon={icon}
+      accessories={ensureCleanAccessories([{ icon: statusIcon }])}
       actions={
         <ActionPanel>
-          <CreateMRAction project={props.project} branch={branch} />
-          <OpenInBrowserAction url={branch.web_url} />
+          <ShowBranchCommitsAction projectID={project.id} branch={branch} />
+          <CreateMRAction project={project} branch={branch} />
+          <GitLabOpenInBrowserAction url={branch.web_url} />
         </ActionPanel>
       }
     />
@@ -43,7 +55,7 @@ export function BranchList(props: { project: Project }) {
   const [query, setQuery] = useState<string>("");
   const { branches, error, isLoading } = useSearch(query, props.project);
   if (error) {
-    showToast(ToastStyle.Failure, "Cannot search branches", error);
+    showToast(Toast.Style.Failure, "Cannot search branches", error);
   }
 
   return (