+++ extensions/homeassistant/src/components/states.tsx
@@ -15,6 +15,8 @@ import {
   SelectSourceAction,
   getMediaPlayerTitleAndArtist,
   CopyTrackToClipboard,
+  MediaPlayerTurnOnAction,
+  MediaPlayerTurnOffAction,
 } from "./mediaplayer";
 import {
   BrightnessControlAction,
@@ -49,6 +51,7 @@ import { InputSelectOptionSelectAction } from "./input_select";
 import { InputButtonPressAction } from "./input_button";
 import { InputTextSetValueAction } from "./input_text";
 import { InputDateTimeSetValueAction } from "./input_datetime";
+import { UpdateInstallAction, UpdateOpenInBrowser, UpdateShowChangelog, UpdateSkipVersionAction } from "./updates";
 
 const PrimaryIconColor = Color.Blue;
 const UnavailableColor = "#bdbdbd";
@@ -120,6 +123,11 @@ function getDeviceClassIcon(state: State): Image.ImageLike | undefined {
       const source = state.state === "on" ? "power-plug.png" : "power-plug-off.png";
       const color = state.state === "unavailable" ? UnavailableColor : PrimaryIconColor;
       return { source: source, tintColor: color };
+    } else if (dc === "door") {
+      const source = state.state === "on" ? "door-open.png" : "door-closed.png";
+      const color =
+        state.state === "unavailable" ? UnavailableColor : state.state === "on" ? Color.Yellow : PrimaryIconColor;
+      return { source: source, tintColor: color };
     }
     const src = deviceClassIconSource[dc] || "entity.png";
     return { source: src, tintColor: PrimaryIconColor };
@@ -153,6 +161,12 @@ function getIcon(state: State): Image.ImageLike | undefined {
     return { source: source, tintColor: color };
   } else if (e.startsWith("person")) {
     return { source: "person.png", tintColor: PrimaryIconColor };
+  } else if (e.startsWith("update")) {
+    const ep = (state.attributes.entity_picture as string) || undefined;
+    if (ep) {
+      return ep.startsWith("/") ? ha.urlJoin(ep) : ep;
+    }
+    return { source: "update.png", tintColor: state.state === "on" ? Color.Yellow : PrimaryIconColor };
   } else if (e.startsWith("cover")) {
     const source = coverStateIconSource[`${state.state}`] || coverStateIconSource.open;
     return { source: source, tintColor: PrimaryIconColor };
@@ -320,11 +334,35 @@ export function StateListItem(props: { state: State }): JSX.Element {
               return "Up To Date";
             }
           }
+        } else if (dc === "door") {
+          switch (state.state) {
+            case "on": {
+              return "Open";
+            }
+            case "off": {
+              return "Closed";
+            }
+          }
         }
       }
       return state.state;
     } else if (state.entity_id.startsWith("input_button")) {
       return new Date(state.state).toISOString().replace("T", " ").replace("Z", "");
+    } else if (state.entity_id.startsWith("update")) {
+      if (state.attributes.in_progress === true) {
+        return "in progress 🔄";
+      }
+      const iv = state.attributes.installed_version;
+      const lv = state.attributes.latest_version;
+      if (state.state === "on" && lv) {
+        if (iv) {
+          return `${iv} => ${lv}`;
+        }
+        return lv;
+      } else if (state.state === "off") {
+        return "✅";
+      }
+      return state.state;
     }
     return state.state;
   };
@@ -507,6 +545,8 @@ export function StateActionPanel(props: { state: State }): JSX.Element {
               onAction={async () => await ha.previousMedia(entityID)}
               icon={{ source: "previous.png", tintColor: Color.PrimaryText }}
             />
+          </ActionPanel.Section>
+          <ActionPanel.Section title="Volume">
             <Action
               title="Volume Up"
               shortcut={{ modifiers: ["cmd"], key: "+" }}
@@ -527,6 +567,10 @@ export function StateActionPanel(props: { state: State }): JSX.Element {
               icon={{ source: Icon.SpeakerSlash, tintColor: Color.PrimaryText }}
             />
             <SelectSourceAction state={state} />
+          </ActionPanel.Section>
+          <ActionPanel.Section title="Power">
+            <MediaPlayerTurnOnAction state={state} />
+            <MediaPlayerTurnOffAction state={state} />
           </ActionPanel.Section>
           <ActionPanel.Section title="Attributes">
             <ShowAttributesAction state={state} />
@@ -956,6 +1000,31 @@ export function StateActionPanel(props: { state: State }): JSX.Element {
         <ActionPanel>
           <ActionPanel.Section title="Controls">
             <InputDateTimeSetValueAction state={state} />
+          </ActionPanel.Section>
+          <ActionPanel.Section title="Attributes">
+            <ShowAttributesAction state={props.state} />
+          </ActionPanel.Section>
+          <ActionPanel.Section title="Values">
+            <CopyEntityIDAction state={state} />
+            <CopyStateValueAction state={state} />
+          </ActionPanel.Section>
+          <ActionPanel.Section title="History">
+            <OpenEntityHistoryAction state={state} />
+            <OpenEntityLogbookAction state={state} />
+          </ActionPanel.Section>
+        </ActionPanel>
+      );
+    }
+    case "update": {
+      return (
+        <ActionPanel>
+          <ActionPanel.Section title="Controls">
+            <UpdateShowChangelog state={state} />
+            <UpdateOpenInBrowser state={state} />
+          </ActionPanel.Section>
+          <ActionPanel.Section title="Install">
+            <UpdateInstallAction state={state} />
+            <UpdateSkipVersionAction state={state} />
           </ActionPanel.Section>
           <ActionPanel.Section title="Attributes">
             <ShowAttributesAction state={props.state} />